refactor: add notes RSS feed + refactor all feeds code
